Move FAH files from init directory to common library as there is no init command (#6945)
